+++ paddle_billing/Entities/Reports/ReportType.py
@@ -10,3 +10,4 @@ class ReportType(PaddleStrEnum, metaclass=PaddleStrEnumMeta):
     TransactionLineItems: "ReportType" = "transaction_line_items"
     Balance: "ReportType" = "balance"
     PayoutReconciliation: "ReportType" = "payout_reconciliation"
+    Checkouts: "ReportType" = "checkouts"