+++ paddle_billing/Resources/CheckoutDomains/Operations/VerifyCheckoutDomainPaymentMethod.py
@@ -1,0 +1,9 @@
+from dataclasses import dataclass
+
+from paddle_billing.Operation import Operation
+from paddle_billing.Entities.CheckoutDomains import CheckoutDomainPaymentMethod
+
+
+@dataclass
+class VerifyCheckoutDomainPaymentMethod(Operation):
+    payment_method: CheckoutDomainPaymentMethod